fix: 分离 react 类型
