+++ example/midway-react-ssr/package.json
@@ -13,7 +13,7 @@
     "react-router-dom": "^5.1.2",
     "serialize-javascript": "^3.0.0",
     "ssr-core-react": "^5.0.0",
-    "ssr-types": "^5.0.0",
+    "ssr-types-react": "^5.0.0",
     "swiper": "^6.0.0"
   },
   "devDependencies": {
