+++ example/nestjs-react-ssr/package.json
@@ -15,7 +15,7 @@
     "rxjs": "^6.6.3",
     "serialize-javascript": "^3.0.0",
     "ssr-core-react": "^5.0.0",
-    "ssr-types": "^5.0.0",
+    "ssr-types-react": "^5.0.0",
     "swiper": "^6.0.0"
   },
   "devDependencies": {
