+++ example/nestjs-react-ssr/web/@types/global.d.ts
@@ -1,4 +1,4 @@
-import { IWindow } from 'ssr-types'
+import { IWindow } from 'ssr-types-react'
 
 declare module '*.less'
