+++ example/nestjs-react-ssr/web/components/layout/App.tsx
@@ -1,7 +1,7 @@
 // 此文件将会在服务端/客户端都将会用到
 // 可通过 __isBrowser__ 或者 useEffect 判断当前在 浏览器环境做一些初始化操作
 
-import { LayoutProps } from 'ssr-types'
+import { LayoutProps } from 'ssr-types-react'
 
 export default (props: LayoutProps) => {
   return props.children!
