+++ example/nestjs-react-ssr/web/pages/index/fetch.ts
@@ -1,4 +1,4 @@
-import { ISSRContext } from 'ssr-types'
+import { ISSRContext } from 'ssr-types-react'
 import { IndexData } from '@/interface'
 interface IApiService {
   index: () => Promise<IndexData>
