+++ example/nestjs-react-ssr/web/pages/index/render.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react'
-import { SProps, IContext } from 'ssr-types'
+import { SProps, IContext } from 'ssr-types-react'
 import Slider from '@/components/slider'
 import Rectangle from '@/components/rectangle'
 import Search from '@/components/search'
