+++ packages/types/src/index.ts
@@ -8,3 +8,10 @@ export type ESMFeRouteItem<T={}> = {
   path: string
   webpackChunkName: string
 } & T
+
+export interface ParseFeRouteItem {
+  path?: string
+  fetch?: string
+  component?: string
+  webpackChunkName?: string
+}
